requirements.txt
openai
anthropic
python-dotenv
requests
# Chess module dependencies
python-chess
torch
streamlit
